requirements.txt
streamlit
requests
python-dotenv
pandas
numpy
google-cloud-bigquery
google-adk
google-generativeai
